backend/requirements.txt
Flask==3.0.0
google-generativeai
beautifulsoup4==4.12.2
selenium
openai==0.28.0
